PLI-47, PLI-46: add auth on BE and FE

+++ libs/plia/editor/api/src/site/site.service.ts
@@ -1,8 +1,10 @@
 import { Repository } from 'typeorm';
-import { Injectable } from '@nestjs/common';
+import { REQUEST } from '@nestjs/core';
+import { Inject, Injectable, Scope } from '@nestjs/common';
 import { InjectRepository } from '@nestjs/typeorm';
 
 import { SiteStatus } from '@plia/plia/types';
+import { UserEntity } from '@plia/plia/user/api';
 
 import { SiteEntity } from './site.entity';
 import { CreateSiteDto } from './dto/create-site.dto';
@@ -10,15 +12,16 @@ import { UpdateSiteDto } from './dto/update-site.dto';
 import { PublishSiteDto } from './dto/publish-site.dto';
 import { HtmlFileService } from './helpers/storage.helper';
 
-@Injectable()
+@Injectable({ scope: Scope.REQUEST })
 export class SiteService {
   constructor(
     @InjectRepository(SiteEntity)
     private siteRepository: Repository<SiteEntity>,
     private htmlFileService: HtmlFileService,
+    @Inject(REQUEST) private request: Request & { user: UserEntity },
   ) {}
 
-  findAll() {
+  findAllByUser() {
     return this.siteRepository.find({
       select: {
         domain: true,
@@ -31,13 +34,18 @@ export class SiteService {
           name: true,
         },
       },
+      where: {
+        user: {
+          id: this.request.user.id,
+        },
+      },
       relations: {
         pages: true,
       },
     });
   }
 
-  findOne(id: string) {
+  findOneByUser(id: string) {
     return this.siteRepository.findOne({
       select: {
         domain: true,
@@ -55,19 +63,36 @@ export class SiteService {
       },
       where: {
         id,
+        user: {
+          id: this.request.user.id,
+        },
       },
     });
   }
 
-  create(site: CreateSiteDto) {
-    return this.siteRepository.save(site);
+  async create(site: CreateSiteDto) {
+    return this.siteRepository.save({
+      ...site,
+      user: {
+        id: this.request.user.id,
+      },
+    });
   }
 
   update(id: string, site: UpdateSiteDto) {
     return this.siteRepository.manager.transaction(async (tsx) => {
-      const prevSite = await this.findOne(id);
+      const prevSite = await this.findOneByUser(id);
 
-      await tsx.update(SiteEntity, { id }, site);
+      await tsx.update(
+        SiteEntity,
+        {
+          id,
+          user: {
+            id: this.request.user.id,
+          },
+        },
+        site,
+      );
 
       if (prevSite.url && prevSite.url !== site.url) {
         await this.htmlFileService.rename(prevSite.url, site.url);
@@ -77,11 +102,18 @@ export class SiteService {
 
   remove(id: string) {
     return this.siteRepository.manager.transaction(async (tsx) => {
-      const site = await this.findOne(id);
+      const site = await this.findOneByUser(id);
 
-      await tsx.delete(SiteEntity, { id });
+      await tsx.delete(SiteEntity, {
+        id,
+        user: {
+          id: this.request.user.id,
+        },
+      });
 
-      await this.htmlFileService.delete(site.url);
+      if (site.url) {
+        await this.htmlFileService.delete(site.url);
+      }
     });
   }
 
@@ -89,7 +121,12 @@ export class SiteService {
     return this.siteRepository.manager.transaction(async (tsx) => {
       await tsx.update(
         SiteEntity,
-        { id },
+        {
+          id,
+          user: {
+            id: this.request.user.id,
+          },
+        },
         {
           name: site.name,
           url: site.url,
